Address feedback and fix false InsecureRequestWarning
A lot of logic was based off of resolve_cert_reqs and resolve_ssl_version.
InsecureRequestWarnings were also being issued based, in part, off of those
values returned. To have an accurate warning, we need to also pass ssl_context
to have an accurate idea of whether the request was secured or not. Further,
we should not override settings on an SSLContext provided by the user. That
said, tests are failing right now that I don't have time to immediately
investigate.